Marged RELENG_1_0 branch changed r1971:2014 into the trunk.
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist/trunk/OpenRTM-aist@2015 8bdb766d-82ee-4cb2-b2de-c55db819e4d5

+++ src/lib/rtm/tests/OutPortCorbaCdrConsumer/OutPortCorbaCdrConsumerTests.cpp
@@ -281,7 +281,7 @@ namespace OutPortCorbaCdrConsumer
 
         RTC::TimedLong rtd;
         rtd <<= indata;
-        CPPUNIT_ASSERT_EQUAL((long)testdata[0], rtd.data);
+        CPPUNIT_ASSERT_EQUAL((long)testdata[0], (long)rtd.data);
 
         CPPUNIT_ASSERT(!CORBA::is_nil(consumer.get_m_objre()));
         consumer.unsubscribeInterface(prof.properties);
